Context-menu bypass working without driver
If the user "KeyBinds" option is disabled, keyboard drivers aren't
loaded.  However, right-click context menu is (or should be!) orthogonal
to keybinds.  This fixes right-click shift bypass when the keybinding is
disabled.

Specifically, the browser context menu swallows the keyup, so this
checks for termination of the browser context menu and then resets the
bypass state.