Remove S.COLLAPSE_ON_PARTIAL_WIN_CLOSE
Don't use the option I started adding in
25752b33290885b25a83bf84ef0b097fee176aea.  It would require winOnRemoved
to receive state information about why the window was closed, and I don't
think this feature is sufficiently valuable to take the memory and
complexity hit.  Let me know if you disagree!

Also, improve the ergonomics of "collapse on window close".  If you hit
Close on an already-closed window, collapse it.

Mentioned in #35.