Port fancy-settings from mootools to jquery
New lib/fancy-settings-jquery.js is my JQuery port of the Mootools
fancy-settings.  I wrote the code from the ground up to match the
behaviour of fancy-settings.  I did not actually try to map Mootools
constructs to JQuery equivalents.

As part of this, I generalized some of the CSS rules so that .searching
does not have to be an attribute of <body>.